update order by date

+++ controllers/users/UserPaymentController.js
@@ -24,9 +24,7 @@ class UserPaymentController {
         });
         for (let i in dataCarts) {
           const package_tripId = dataCarts[i].package_tripId;
-          const dataPack = await package_trip.findOne({
-            where: { id: package_tripId },
-          });
+          const dataPack = await package_trip.findOne({ where: { id: package_tripId } });
           const images = await temp_image.findAll({
             attributes: ["id", "package_tripId", "img"],
             where: { package_tripId },
@@ -63,9 +61,7 @@ class UserPaymentController {
         let isBooked = false;
         today.setHours(0, 0, 0, 0);
         if (varDate >= today) {
-          const valCart = await cart_item.findAll({
-            where: { date, package_tripId },
-          });
+          const valCart = await cart_item.findAll({ where: { date, package_tripId } });
           for (let i in valCart) {
             let paymentId = valCart[i].paymentId;
             const valP = await payment.findOne({ where: { id: paymentId } });
@@ -75,40 +71,26 @@ class UserPaymentController {
             }
           }
           if (isBooked) {
-            res.status(203).json({
-              msg: "This package trip already booked by someone, choose another date!",
-            });
+            res.status(203).json({ msg: "This package trip already booked by someone, choose another date!" });
           } else {
-            const valPackage = await package_trip.findOne({
-              where: { id: package_tripId },
-            });
+            const valPackage = await package_trip.findOne({ where: { id: package_tripId } });
             if (valPackage) {
-              const valPaym = await payment.findOne({
-                where: { userId, status: "oncart" },
-              });
+              const valPaym = await payment.findOne({ where: { userId, status: "oncart" } });
               const price = +valPackage.price;
               if (valPaym) {
-                const valSameData = await cart_item.findOne({
-                  where: { date, package_tripId, paymentId: valPaym.id },
-                });
+                const valSameData = await cart_item.findOne({ where: { date, package_tripId, paymentId: valPaym.id } });
                 if (valSameData) {
                   const newAmount = +valSameData.amount + +amount;
                   if (newAmount > 8) {
                     res.status(203).json({ msg: "Max member has reach out!" });
                   } else {
                     await cart_item.update({ amount: newAmount }, { where: { id: valSameData.id } });
-                    const result = await cart_item.findOne({
-                      where: { id: valSameData.id },
-                    });
+                    const result = await cart_item.findOne({ where: { id: valSameData.id } });
                     let balance = 0;
-                    const dataTotal = await cart_item.findAll({
-                      where: { paymentId: +valPaym.id },
-                    });
+                    const dataTotal = await cart_item.findAll({ where: { paymentId: +valPaym.id } });
                     for (let i in dataTotal) {
                       const id = +dataTotal[i].package_tripId;
-                      const { price } = await package_trip.findOne({
-                        where: { id },
-                      });
+                      const { price } = await package_trip.findOne({ where: { id } });
                       let jumlah = price * +dataTotal[i].amount;
                       balance += jumlah;
                     }
@@ -116,21 +98,12 @@ class UserPaymentController {
                     res.status(201).json(result);
                   }
                 } else {
-                  const addCart = await cart_item.create({
-                    paymentId: valPaym.id,
-                    date,
-                    package_tripId,
-                    amount,
-                  });
+                  const addCart = await cart_item.create({ paymentId: valPaym.id, date, package_tripId, amount });
                   let balance = 0;
-                  const dataTotal = await cart_item.findAll({
-                    where: { paymentId: valPaym.id },
-                  });
+                  const dataTotal = await cart_item.findAll({ where: { paymentId: valPaym.id } });
                   for (let i in dataTotal) {
                     const id = dataTotal[i].package_tripId;
-                    const { price } = await package_trip.findOne({
-                      where: { id },
-                    });
+                    const { price } = await package_trip.findOne({ where: { id } });
                     let jumlah = price * dataTotal[i].amount;
                     balance += jumlah;
                   }
@@ -138,17 +111,9 @@ class UserPaymentController {
                   res.status(201).json(addCart);
                 }
               } else {
-                const newPay = await payment.create({
-                  userId,
-                  status: "oncart",
-                });
+                const newPay = await payment.create({ userId, status: "oncart" });
                 const paymentId = newPay.id;
-                const addCart = await cart_item.create({
-                  paymentId,
-                  date,
-                  package_tripId,
-                  amount,
-                });
+                const addCart = await cart_item.create({ paymentId, date, package_tripId, amount });
                 const total = price * +addCart.amount;
                 await payment.update({ total }, { where: { id: newPay.id } });
                 res.status(201).json(addCart);
@@ -172,13 +137,9 @@ class UserPaymentController {
       const { id } = req.body;
       //tambah validasi payment code : nanti
       const payment_code = (Math.random() + 1).toString(36).substring(2).toUpperCase();
-      const valPaym = await payment.findOne({
-        where: { userId, status: "oncart" },
-      });
+      const valPaym = await payment.findOne({ where: { userId, status: "oncart" } });
       if (valPaym) {
-        const validasiCart = await cart_item.findAll({
-          where: { id, paymentId: valPaym.id },
-        });
+        const validasiCart = await cart_item.findAll({ where: { id, paymentId: valPaym.id } });
         let isValidDate = true;
         let today = new Date();
         for (let j in validasiCart) {
@@ -190,9 +151,7 @@ class UserPaymentController {
           }
         }
         if (isValidDate) {
-          const listCarts = await cart_item.findAll({
-            where: { paymentId: valPaym.id },
-          });
+          const listCarts = await cart_item.findAll({ where: { paymentId: valPaym.id } });
           const arrCarts = [];
           listCarts.forEach((list) => {
             return arrCarts.push(list.id);
@@ -212,16 +171,10 @@ class UserPaymentController {
             }
           }
           if (valCart) {
-            const newPayment = await payment.create({
-              userId,
-              payment_code,
-              status: "unpaid",
-            });
+            const newPayment = await payment.create({ userId, payment_code, status: "unpaid" });
             await cart_item.update({ paymentId: newPayment.id }, { where: { id } });
             let balance = 0;
-            const dataTotal = await cart_item.findAll({
-              where: { paymentId: valPaym.id },
-            });
+            const dataTotal = await cart_item.findAll({ where: { paymentId: valPaym.id } });
             for (let i in dataTotal) {
               const id = dataTotal[i].package_tripId;
               const { price } = await package_trip.findOne({ where: { id } });
@@ -230,9 +183,7 @@ class UserPaymentController {
             }
             await payment.update({ total: balance }, { where: { id: valPaym.id } });
             let totalUnpaid = 0;
-            const dataTotalUnpaid = await cart_item.findAll({
-              where: { paymentId: newPayment.id },
-            });
+            const dataTotalUnpaid = await cart_item.findAll({ where: { paymentId: newPayment.id } });
             for (let i in dataTotalUnpaid) {
               const id = dataTotalUnpaid[i].package_tripId;
               const { price } = await package_trip.findOne({ where: { id } });
@@ -240,9 +191,7 @@ class UserPaymentController {
               totalUnpaid += jumlah;
             }
             await payment.update({ total: totalUnpaid }, { where: { id: newPayment.id } });
-            const dataCarts = await cart_item.findAll({
-              where: { paymentId: newPayment.id },
-            });
+            const dataCarts = await cart_item.findAll({ where: { paymentId: newPayment.id } });
             res.json(dataCarts);
           } else {
             res.status(404).json({ msg: "Id cart_item isnt match!" });
@@ -263,8 +212,9 @@ class UserPaymentController {
       const userId = req.user.id;
       const result = [];
       const valPaym = await payment.findAll({
-        attributes: { exclude: ["createdAt", "updatedAt"] },
+        attributes: { exclude: ["createdAt"] },
         where: { userId, status: { [Op.or]: ["pending", "unpaid"] } },
+        order: [["updatedAt", "DESC"]],
       });
       if (valPaym) {
         for (let a in valPaym) {
@@ -275,9 +225,7 @@ class UserPaymentController {
           let temp = [];
           for (let i in dataCarts) {
             const package_tripId = dataCarts[i].package_tripId;
-            const dataPack = await package_trip.findOne({
-              where: { id: package_tripId },
-            });
+            const dataPack = await package_trip.findOne({ where: { id: package_tripId } });
             const images = await temp_image.findAll({
               attributes: ["id", "package_tripId", "img"],
               where: { package_tripId },
@@ -306,8 +254,9 @@ class UserPaymentController {
       const userId = req.user.id;
       const result = [];
       const valPaym = await payment.findAll({
-        attributes: { exclude: ["createdAt", "updatedAt"] },
+        attributes: { exclude: ["createdAt"] },
         where: { userId, status: "paid" },
+        order: [["updatedAt", "DESC"]],
       });
       if (valPaym) {
         for (let a in valPaym) {
@@ -318,9 +267,7 @@ class UserPaymentController {
           let temp = [];
           for (let i in dataCarts) {
             const package_tripId = dataCarts[i].package_tripId;
-            const dataPack = await package_trip.findOne({
-              where: { id: package_tripId },
-            });
+            const dataPack = await package_trip.findOne({ where: { id: package_tripId } });
             const images = await temp_image.findAll({
               attributes: ["id", "package_tripId", "img"],
               where: { package_tripId },
@@ -348,9 +295,7 @@ class UserPaymentController {
     try {
       const id = +req.params.id;
       const userId = req.user.id;
-      const result = await payment.destroy({
-        where: { id, userId, status: "unpaid" },
-      });
+      const result = await payment.destroy({ where: { id, userId, status: "unpaid" } });
       if (result !== 0) {
         res.status(200).json({ message: `Payment with id ${id} has been deleted` });
       } else {
@@ -367,15 +312,11 @@ class UserPaymentController {
       const userId = req.user.id;
       const valCart = await cart_item.findOne({ where: { id } });
       if (valCart) {
-        const valPayment = await payment.findOne({
-          where: { id: valCart.paymentId, userId, status: "oncart" },
-        });
+        const valPayment = await payment.findOne({ where: { id: valCart.paymentId, userId, status: "oncart" } });
         if (valPayment) {
           await cart_item.destroy({ where: { id } });
           let balance = 0;
-          const dataTotal = await cart_item.findAll({
-            where: { paymentId: +valPayment.id },
-          });
+          const dataTotal = await cart_item.findAll({ where: { paymentId: +valPayment.id } });
           for (let i in dataTotal) {
             const id = +dataTotal[i].package_tripId;
             const { price } = await package_trip.findOne({ where: { id } });
@@ -400,18 +341,12 @@ class UserPaymentController {
       const id = +req.params.id;
       const userId = req.user.id;
       let isBooked = false;
-      const cekPayment = await payment.findOne({
-        where: { id, userId, status: "unpaid" },
-      });
+      const cekPayment = await payment.findOne({ where: { id, userId, status: "unpaid" } });
       if (cekPayment) {
-        const valPayment = await cart_item.findAll({
-          where: { paymentId: id },
-        });
+        const valPayment = await cart_item.findAll({ where: { paymentId: id } });
         for (let a in valPayment) {
           const { date, package_tripId } = valPayment[a];
-          const valCart = await cart_item.findAll({
-            where: { date, package_tripId },
-          });
+          const valCart = await cart_item.findAll({ where: { date, package_tripId } });
           for (let i in valCart) {
             let paymentId = valCart[i].paymentId;
             const valP = await payment.findOne({ where: { id: paymentId } });
@@ -422,9 +357,7 @@ class UserPaymentController {
           }
         }
         if (isBooked) {
-          res.status(203).json({
-            msg: "This package trip already booked by someone, choose another date!",
-          });
+          res.status(203).json({ msg: "This package trip already booked by someone, choose another date!" });
         } else {
           const result = await payment.update({ status: "paid" }, { where: { id, userId, status: "unpaid" } });
           if (result[0] !== 0) {
@@ -446,9 +379,7 @@ class UserPaymentController {
       // TODO tambah validasi
       const id = +req.params.id;
       const userId = +req.user.id;
-      const { payment_code, total } = await payment.findOne({
-        where: { id, userId },
-      });
+      const { payment_code, total } = await payment.findOne({ where: { id, userId } });
       const { payment_type, bank } = req.body;
 
       const payload = {
@@ -479,7 +410,7 @@ class UserPaymentController {
 
   static async notification(req, res) {
     try {
-      const notif = await core.transaction.notification(req.body);
+      const notif = await core.transaction.notification(notificationJson);
       const code = notif.order_id;
       let status = "";
       notif.transaction_status === "settlement" ? (status = "paid") : (status = notif.transaction_status);
@@ -505,7 +436,7 @@ class UserPaymentController {
       if (status === "paid") {
         res.status(200).json(result);
       } else {
-        res.status(400).json({ msg: "Please complete your transaction." });
+        res.status(200).json({ msg: "Please complete your transaction." });
       }
     } catch (error) {
       res.status(500).json(error);
@@ -517,23 +448,12 @@ class UserPaymentController {
       const id = +req.params.id;
       const userId = +req.user.id;
       let result = {};
-      const dataPayment = await payment.findOne({
-        where: { id, userId, status: "pending" },
-      });
+      const dataPayment = await payment.findOne({ where: { id, userId, status: "pending" } });
       const { payment_code, total, status } = dataPayment;
       const responseMidtrans = JSON.parse(dataPayment.responseMidtrans);
       const { va_numbers, payment_type } = responseMidtrans;
       const { bank, va_number } = va_numbers[0];
-      result = {
-        id,
-        userId,
-        payment_code,
-        total,
-        payment_type,
-        bank,
-        va_number,
-        status,
-      };
+      result = { id, userId, payment_code, total, payment_type, bank, va_number, status };
       res.status(200).json(result);
     } catch (error) {
       res.status(500).json(error);
